chore: convert more tests
