--- packages/core/admin/admin/src/pages/MarketplacePage/tests/mocks/providers.js
@@ -1,560 +1,0 @@
-const providers = {
-  data: [
-    {
-      id: 'rec6lwImIw1HvvE5X',
-      attributes: {
-        name: 'Amazon Ses',
-        description: 'Amazon Ses provider for Strapi',
-        slug: '@strapi-provider-email-amazon-ses',
-        npmPackageName: '@strapi/provider-email-amazon-ses',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-email-amazon-ses',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/email-amazon-ses',
-        pluginName: 'Email',
-        logo: {
-          id: 'attPkleJsAHaDv8vC',
-          width: 256,
-          height: 256,
-          url: 'https://dl.airtable.com/.attachments/04be9ade6077d029a9d4108cd2f47c8e/f5fe8d67/amazonSES.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=0561c26f91b25e21',
-          filename: 'amazonSES.png',
-          size: 17284,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/c9643af6593cacd432a65755c6cc8369/bd74f743?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=15a41c58b4726d77',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/ae318dfd252d61d4c6302e055e883df0/aa87e132?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=9277ec20b0fbef11',
-              width: 256,
-              height: 256,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/4bf5b25a4f4be4fc3da2a145406b0aa1/095a0e80?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=74744934a5c587e5',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-      },
-    },
-    {
-      id: 'recaK5lCAUko6TJ6l',
-      attributes: {
-        name: 'AWS S3',
-        description: 'AWS S3 provider for Strapi uploads',
-        slug: '@strapi-provider-upload-aws-s3',
-        npmPackageName: '@strapi/provider-upload-aws-s3',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-upload-aws-s3',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/upload-aws-s3',
-        pluginName: 'Upload',
-        logo: {
-          id: 'att0O0zMQjUcgO9An',
-          width: 428,
-          height: 512,
-          url: 'https://dl.airtable.com/.attachments/e24c0ef99b1e952cbf0c8cbeed8cc24f/da30ba27/Amazon-S3-Logo.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=953eaf833eba09a2',
-          filename: 'Amazon-S3-Logo.png',
-          size: 8747,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/c321c1a3812cfe2fec89d50e59608dfc/f33dfcb7?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=aa69cd4276a975c5',
-              width: 30,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/9ea423169f9f584d764bf3dc4309f27c/acb36ad2?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=787909ce6eb1c12f',
-              width: 428,
-              height: 512,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/3cd1f6a2d6470f4fd6f241a4fd49863e/46d55428?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=e7a579bb13a12962',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-        strapiCompatibility: 'v4',
-      },
-    },
-    {
-      id: 'recQCFBl61THeLN0w',
-      attributes: {
-        name: 'Cloudinary',
-        description: 'Cloudinary provider for Strapi uploads',
-        slug: '@strapi-provider-upload-cloudinary',
-        npmPackageName: '@strapi/provider-upload-cloudinary',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-upload-cloudinary',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/upload-cloudinary',
-        pluginName: 'Upload',
-        logo: {
-          id: 'attWcLyN8uXzKMkyp',
-          width: 1000,
-          height: 1000,
-          url: 'https://dl.airtable.com/.attachments/74df39c3715a78589be1cbff69009dc2/14734e59/cloudinary.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=d82f74e2c1563864',
-          filename: 'cloudinary.png',
-          size: 40035,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/09e31f64effadbb0db337ce62ce6e82f/41c381e5?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=009f1b66ad3ce551',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/73a94b8533235b8b16c35971555c4c7c/2ff7c5de?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=f8f69638a0122a41',
-              width: 512,
-              height: 512,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/5f5f4c13f8dddeaffd790c91c01f2de1/3f9fe205?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=3ccfc4325d8c2950',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-        strapiCompatibility: 'v4',
-        githubStars: 49130,
-        npmDownloads: 7492,
-      },
-    },
-    {
-      id: 'recnzqXkRQ0y69gd4',
-      attributes: {
-        name: 'Local Upload',
-        description: 'Local upload provider for Strapi',
-        slug: '@strapi-provider-upload-local',
-        npmPackageName: '@strapi/provider-upload-local',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-upload-local',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/upload-local',
-        pluginName: 'Upload',
-        logo: {
-          id: 'att8OCweGsCtYGZeR',
-          width: 80,
-          height: 80,
-          url: 'https://dl.airtable.com/.attachments/6707fcab8f5530214160bacf6f4369ec/9ed22aaf/typeplaceholderversion1.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=35d9e0bb75e4528a',
-          filename: 'type=placeholder, version=1.png',
-          size: 1460,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/8d7f33de80df88b2a2c0bd0b655a8cb5/87df3cd7?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=84f984cef71ffa27',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/5b047bcf648bcb742b604ca69e2f668d/90d7caa8?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=16607a5798ba4057',
-              width: 80,
-              height: 80,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/499515fba624161eff1042e9bb7c1e5e/ecae1616?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=eb01b0394a8c1c38',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-        githubStars: 30,
-        npmDownloads: 2492,
-      },
-    },
-    {
-      id: 'rec65qlVCX8mD95Uy',
-      attributes: {
-        name: 'Mailgun',
-        description: 'Mailgun provider for Strapi',
-        slug: '@strapi-provider-email-mailgun',
-        npmPackageName: '@strapi/provider-email-mailgun',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-email-mailgun',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/email-mailgun',
-        pluginName: 'Email',
-        logo: {
-          id: 'attXAPE3yfQFOiZM0',
-          width: 299,
-          height: 300,
-          url: 'https://dl.airtable.com/.attachments/73118ab0aea4d9eae66edc18a1db4982/808842ba/mailgun-logo-5388F66106-seeklogo_com.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=2e2e9ecb19e752ca',
-          filename: 'mailgun-logo-5388F66106-seeklogo_com.png',
-          size: 19931,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/70bdc0cc08226c5f54606f1b112ecc89/d57aeb29?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=1796f030bacf7ccf',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/bca142748177bbcea58303c80461f3db/dce79b36?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=8b26f9319deb2fe3',
-              width: 299,
-              height: 300,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/f1978a5f2fc19ab02d1cb5e987481661/6389d07c?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=c4d70a9a26baf3b1',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-      },
-    },
-    {
-      id: 'recIPiVtySXJI0xGc',
-      attributes: {
-        name: 'Nodemailer',
-        description: 'Nodemailer provider for Strapi',
-        slug: '@strapi-provider-email-nodemailer',
-        npmPackageName: '@strapi/provider-email-nodemailer',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-email-nodemailer',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/email-nodemailer',
-        pluginName: 'Email',
-        logo: {
-          id: 'attu6UsiH02uW4mwL',
-          width: 200,
-          height: 200,
-          url: 'https://dl.airtable.com/.attachments/73bea518f68fc312ec15a4a988da3bbc/4c024fea/Nodemailer.jpeg?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=a5dec98be44fa4e6',
-          filename: 'Nodemailer.jpeg',
-          size: 5055,
-          type: 'image/jpeg',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/9cc0b414f2653afb4286238c8ec522fe/a914a2b0?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=e370f43cd0e9b772',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/ab632a86e17718e9d98b10ccb926be75/640eb57d?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=9479daaa35f0751c',
-              width: 200,
-              height: 200,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/14c63b4b32d5d4efb846705ad6a67c34/0310e71e?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=d6ed6efdb841be01',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-      },
-    },
-    {
-      id: 'recL9LW4nothrsZhh',
-      attributes: {
-        name: 'Rackspace',
-        description: 'Rackspace provider for Strapi uploads',
-        slug: '@strapi-provider-upload-rackspace',
-        npmPackageName: '@strapi/provider-upload-rackspace',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-upload-rackspace',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/upload-rackspace',
-        pluginName: 'Upload',
-        logo: {
-          id: 'attwjN4k8kRJqlP7q',
-          width: 1024,
-          height: 1024,
-          url: 'https://dl.airtable.com/.attachments/f1e20497044cc6035f43b94f46bd5381/c78cbd9f/rackspace-logo.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=641682ec9adc2d1c',
-          filename: 'rackspace-logo.png',
-          size: 18186,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/8f9bed56940dbbf48d9e0d137b7ef725/1ddefd5c?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=e8a6fa209cf90d7d',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/fba1cb15c6f10602abbf43c8017cab0d/f1b221a8?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=84e391b031a95da3',
-              width: 512,
-              height: 512,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/7eb59048226a755cdaf03f430c634293/b85dd69d?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=04b1f6ac2595a233',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-        strapiCompatibility: 'v4',
-      },
-    },
-    {
-      id: 'recuuU81JF0lbc0Z3',
-      attributes: {
-        name: 'SendGrid',
-        description: 'SendGrid provider for Strapi',
-        slug: '@strapi-provider-email-sendgrid',
-        npmPackageName: '@strapi/provider-email-sendgrid',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-email-sendgrid',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/email-sendgrid',
-        pluginName: 'Email',
-        logo: {
-          id: 'att3OLz4K78pUVTjt',
-          width: 300,
-          height: 300,
-          url: 'https://dl.airtable.com/.attachments/c09bfc173321ee79a972a10735ed6cf7/4e0a5321/sendgrid-logo-7574E52082-seeklogo_com.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=1ff0a128488e143f',
-          filename: 'sendgrid-logo-7574E52082-seeklogo_com.png',
-          size: 3124,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/7f9a60bffd359b890d697f6a4c713854/d0f3170f?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=2be1006f6571a145',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/36018ddc97bc30cfb2c38dae7dbd39d0/87a95a5e?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=7e86b0c57901541a',
-              width: 300,
-              height: 300,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/d6193a48fe30fa59e84ffc5a400c731f/199e4af5?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=4d7e9d25cfcd3eee',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-      },
-    },
-    {
-      id: 'recG4pjF6Dg8hEG6z',
-      attributes: {
-        name: 'Sendmail',
-        description: 'Sendmail provider for Strapi',
-        slug: '@strapi-provider-email-sendmail',
-        npmPackageName: '@strapi/provider-email-sendmail',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-email-sendmail',
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/email-sendmail',
-        pluginName: 'Email',
-        logo: {
-          id: 'att5AVRI4yF9iI5vr',
-          width: 2400,
-          height: 2400,
-          url: 'https://dl.airtable.com/.attachments/0f14c78742105dffd36d6e253612f992/2f0e8605/sendmail-logo-png-transparent.png?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=65d4bd3786c6b0a2',
-          filename: 'sendmail-logo-png-transparent.png',
-          size: 89691,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/9fed2bb13d53fe0cd1d770c5af83ff07/dcd12795?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=853034a7a274633f',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/264a135d3364255e5e97cd699632fbe9/83835476?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=e515faafd9024636',
-              width: 512,
-              height: 512,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/7d89484dad9a056c482f869b7b3db0d2/be86bd9c?ts=1654603863&userId=usrUa8HWbsGCCzcQm&cs=8469840a78f8719d',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-      },
-    },
-  ],
-  meta: {
-    collections: {
-      'Made by official partners': 0,
-      'Made by Strapi': 6,
-      'Made by the community': 2,
-      Verified: 6,
-    },
-    pagination: {
-      page: 1,
-      pageSize: 24,
-      pageCount: 4,
-      total: 96,
-    },
-  },
-};
-
-const madeByStrapiResponseData = {
-  data: [
-    {
-      id: 'rec6lwImIw1HvvE5X',
-      attributes: {
-        name: 'Amazon SES',
-        description: 'Amazon SES provider for Strapi',
-        slug: '@strapi-provider-email-amazon-ses',
-        npmPackageName: '@strapi/provider-email-amazon-ses',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-email-amazon-ses',
-        npmDownloads: 2070,
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/email-amazon-ses',
-        githubStars: 49137,
-        pluginName: 'Email',
-        logo: {
-          id: 'attPkleJsAHaDv8vC',
-          width: 256,
-          height: 256,
-          url: 'https://dl.airtable.com/.attachments/04be9ade6077d029a9d4108cd2f47c8e/f5fe8d67/amazonSES.png',
-          filename: 'amazonSES.png',
-          size: 17284,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/c9643af6593cacd432a65755c6cc8369/bd74f743',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/ae318dfd252d61d4c6302e055e883df0/aa87e132',
-              width: 256,
-              height: 256,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/4bf5b25a4f4be4fc3da2a145406b0aa1/095a0e80',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-        submissionDate: '2022-05-18T16:29:00.000Z',
-        collections: ['Verified', 'Made by Strapi'],
-      },
-    },
-    {
-      id: 'recaK5lCAUko6TJ6l',
-      attributes: {
-        name: 'AWS S3',
-        description: 'AWS S3 provider for Strapi uploads',
-        slug: '@strapi-provider-upload-aws-s3',
-        npmPackageName: '@strapi/provider-upload-aws-s3',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-upload-aws-s3',
-        npmDownloads: 12088,
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/upload-aws-s3',
-        githubStars: 49137,
-        pluginName: 'Upload',
-        logo: {
-          id: 'att0O0zMQjUcgO9An',
-          width: 428,
-          height: 512,
-          url: 'https://dl.airtable.com/.attachments/e24c0ef99b1e952cbf0c8cbeed8cc24f/da30ba27/Amazon-S3-Logo.png',
-          filename: 'Amazon-S3-Logo.png',
-          size: 8747,
-          type: 'image/png',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/c321c1a3812cfe2fec89d50e59608dfc/f33dfcb7',
-              width: 30,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/9ea423169f9f584d764bf3dc4309f27c/acb36ad2',
-              width: 428,
-              height: 512,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/3cd1f6a2d6470f4fd6f241a4fd49863e/46d55428',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-        strapiCompatibility: 'v4',
-        submissionDate: '2022-05-17T09:07:27.000Z',
-        collections: ['Verified', 'Made by Strapi'],
-      },
-    },
-  ],
-  meta: providers.meta,
-};
-
-const verifiedResponseData = {
-  data: [
-    {
-      id: 'recIPiVtySXJI0xGc',
-      attributes: {
-        name: 'Nodemailer',
-        description: 'Nodemailer provider for Strapi',
-        slug: '@strapi-provider-email-nodemailer',
-        npmPackageName: '@strapi/provider-email-nodemailer',
-        npmPackageUrl: 'https://www.npmjs.com/package/@strapi/provider-email-nodemailer',
-        npmDownloads: 3723,
-        repositoryUrl:
-          'https://github.com/strapi/strapi/tree/master/packages/providers/email-nodemailer',
-        githubStars: 49137,
-        pluginName: 'Email',
-        logo: {
-          id: 'attu6UsiH02uW4mwL',
-          width: 200,
-          height: 200,
-          url: 'https://dl.airtable.com/.attachments/73bea518f68fc312ec15a4a988da3bbc/4c024fea/Nodemailer.jpeg',
-          filename: 'Nodemailer.jpeg',
-          size: 5055,
-          type: 'image/jpeg',
-          thumbnails: {
-            small: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/9cc0b414f2653afb4286238c8ec522fe/a914a2b0',
-              width: 36,
-              height: 36,
-            },
-            large: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/ab632a86e17718e9d98b10ccb926be75/640eb57d',
-              width: 200,
-              height: 200,
-            },
-            full: {
-              url: 'https://dl.airtable.com/.attachmentThumbnails/14c63b4b32d5d4efb846705ad6a67c34/0310e71e',
-              width: 3000,
-              height: 3000,
-            },
-          },
-        },
-        developerName: 'Strapi team',
-        validated: true,
-        madeByStrapi: true,
-        submissionDate: '2022-05-19T07:34:40.000Z',
-        collections: ['Verified'],
-      },
-    },
-  ],
-  meta: providers.meta,
-};
-
-export const responses = {
-  providers,
-  'Made by Strapi': madeByStrapiResponseData,
-  Verified: verifiedResponseData,
-};
