+++ packages/core/admin/admin/src/pages/MarketplacePage/tests/providers.test.js
@@ -1,16 +1,10 @@
 import React from 'react';
 
-import { lightTheme, ThemeProvider } from '@strapi/design-system';
-import { NotificationsProvider } from '@strapi/helper-plugin';
-import { render as renderRTL, screen, waitFor, within } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import { IntlProvider } from 'react-intl';
-import { QueryClient, QueryClientProvider } from 'react-query';
-import { MemoryRouter, Route } from 'react-router-dom';
+import { screen, within } from '@testing-library/react';
+import { render as renderRTL, waitFor } from '@tests/utils';
+import { Route } from 'react-router-dom';
 
 import { MarketPlacePage } from '../index';
-
-import server from './server';
 
 // Increase the jest timeout to accommodate long running tests
 jest.setTimeout(50000);
@@ -22,6 +16,9 @@ jest.mock('../../../hooks/useDebounce', () => ({
   useDebounce: jest.fn((value) => value),
 }));
 jest.mock('../hooks/useNavigatorOnline');
+/**
+ * TODO: remove this mock.
+ */
 jest.mock('@strapi/helper-plugin', () => ({
   ...jest.requireActual('@strapi/helper-plugin'),
   useAppInfo: jest.fn(() => ({
@@ -36,57 +33,33 @@ jest.mock('@strapi/helper-plugin', () => ({
 
 let testLocation = null;
 
-const render = (props) => ({
-  ...renderRTL(<MarketPlacePage {...props} />, {
-    wrapper({ children }) {
-      const client = new QueryClient({
-        defaultOptions: {
-          queries: {
-            retry: false,
-          },
-        },
-      });
+const render = (props) =>
+  renderRTL(<MarketPlacePage {...props} />, {
+    renderOptions: {
+      wrapper({ children }) {
+        return (
+          <>
+            {children}
+            <Route
+              path="*"
+              render={({ location }) => {
+                testLocation = location;
 
-      return (
-        <QueryClientProvider client={client}>
-          <IntlProvider locale="en" messages={{}} textComponent="span">
-            <ThemeProvider theme={lightTheme}>
-              <NotificationsProvider>
-                <MemoryRouter initialEntries={['/?npmPackageType=provider&sort=name:asc']}>
-                  {children}
-                  <Route
-                    path="*"
-                    render={({ location }) => {
-                      testLocation = location;
-
-                      return null;
-                    }}
-                  />
-                </MemoryRouter>
-              </NotificationsProvider>
-            </ThemeProvider>
-          </IntlProvider>
-        </QueryClientProvider>
-      );
+                return null;
+              }}
+            />
+          </>
+        );
+      },
     },
-  }),
-
-  user: userEvent.setup(),
-});
+    initialEntries: ['/?npmPackageType=provider&sort=name:asc'],
+  });
 
 const waitForReload = async () => {
   await waitFor(() => expect(screen.queryByText('Loading content...')).not.toBeInTheDocument());
 };
 
 describe('Marketplace page - providers tab', () => {
-  beforeAll(() => server.listen());
-
-  afterEach(() => {
-    server.resetHandlers();
-  });
-
-  afterAll(() => server.close());
-
   it('renders the providers tab', async () => {
     const { getByText, getByRole, queryByText } = render();
