+++ packages/core/admin/admin/src/pages/ProfilePage/tests/index.test.js
@@ -1,86 +1,53 @@
 import React from 'react';
 
-import { darkTheme, lightTheme } from '@strapi/design-system';
-import { render, screen, waitFor } from '@testing-library/react';
-import { IntlProvider } from 'react-intl';
-import { QueryClient, QueryClientProvider } from 'react-query';
-import { MemoryRouter } from 'react-router-dom';
+import { render, waitFor, server } from '@tests/utils';
+import { rest } from 'msw';
 
-import { Theme } from '../../../components/Theme';
-import { ThemeToggleProvider } from '../../../components/ThemeToggleProvider';
 import ProfilePage from '../index';
-
-import server from './utils/server';
-import serverLockedSSO from './utils/serverLockedSSO';
 
 jest.mock('../../../components/LocalesProvider/useLocalesProvider');
 
 jest.mock('@strapi/helper-plugin', () => ({
   ...jest.requireActual('@strapi/helper-plugin'),
-  useNotification: jest.fn(),
   useFocusWhenNavigate: jest.fn(),
   useAppInfo: jest.fn(() => ({ setUserDisplayName: jest.fn() })),
   useOverlayBlocker: jest.fn(() => ({ lockApp: jest.fn, unlockApp: jest.fn() })),
 }));
 
-const setup = (props) =>
-  render(<ProfilePage {...props} />, {
-    wrapper({ children }) {
-      window.strapi.isEE = true;
-      window.strapi.features.isEnabled = () => true;
-      const client = new QueryClient({
-        defaultOptions: {
-          queries: {
-            retry: false,
-          },
-        },
-      });
+describe('Profile page', () => {
+  const originalIsEnabled = window.strapi.features.isEnabled;
+  const originalIsEE = window.strapi.isEE;
 
-      return (
-        <QueryClientProvider client={client}>
-          <IntlProvider messages={{}} textComponent="span" locale="en">
-            <ThemeToggleProvider themes={{ light: lightTheme, dark: darkTheme }}>
-              <MemoryRouter>
-                <Theme>{children}</Theme>
-              </MemoryRouter>
-            </ThemeToggleProvider>
-          </IntlProvider>
-        </QueryClientProvider>
-      );
-    },
+  beforeAll(() => {
+    window.strapi.isEE = true;
+    window.strapi.features.isEnabled = () => true;
   });
 
-describe('ADMIN | Pages | Profile page | without SSO lock', () => {
-  beforeAll(() => server.listen());
+  afterAll(() => {
+    window.strapi.isEE = originalIsEE;
+    window.strapi.features.isEnabled = originalIsEnabled;
+  });
 
   beforeEach(() => {
     jest.clearAllMocks();
   });
 
-  afterEach(() => {
-    server.resetHandlers();
-  });
-
-  afterAll(() => {
-    server.close();
-  });
-
   it('renders and show the Interface Language section', async () => {
-    setup();
+    const { getByText } = render(<ProfilePage />);
     await waitFor(() => {
-      expect(screen.getByText('Interface language')).toBeInTheDocument();
+      expect(getByText('Interface language')).toBeInTheDocument();
     });
   });
 
   it('should display username if it exists', async () => {
-    setup();
+    const { getByText } = render(<ProfilePage />);
     await waitFor(() => {
-      expect(screen.getByText('yolo')).toBeInTheDocument();
+      expect(getByText('yolo')).toBeInTheDocument();
     });
   });
 
   it('should display the change password section and all its fields', async () => {
-    const { getByRole, queryByTestId, getByLabelText } = setup();
+    const { getByRole, queryByTestId, getByLabelText } = render(<ProfilePage />);
 
     await waitFor(() => {
       expect(queryByTestId('loader')).not.toBeInTheDocument();
@@ -98,33 +65,22 @@ describe('ADMIN | Pages | Profile page | without SSO lock', () => {
 
     expect(getByLabelText('Password confirmation')).toBeInTheDocument();
   });
-});
-
-describe('ADMIN | Pages | Profile page | with SSO lock', () => {
-  beforeAll(() => serverLockedSSO.listen());
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  afterEach(() => {
-    serverLockedSSO.resetHandlers();
-  });
-
-  afterAll(() => {
-    jest.clearAllMocks();
-    serverLockedSSO.close();
-  });
-
-  it('should display username if it exists', async () => {
-    setup();
-    await waitFor(() => {
-      expect(screen.getByText('yolo')).toBeInTheDocument();
-    });
-  });
 
   it('should not display the change password section and all the fields if the user role is Locked', async () => {
-    const { queryByRole, queryByTestId, queryByLabelText } = setup();
+    server.use(
+      rest.get('/admin/providers/isSSOLocked', (req, res, ctx) => {
+        return res(
+          ctx.json({
+            data: {
+              isSSOLocked: true,
+            },
+          })
+        );
+      })
+    );
+
+    const { queryByRole, queryByTestId, queryByLabelText } = render(<ProfilePage />);
+
     const changePasswordHeading = queryByRole('heading', {
       name: 'Change password',
     });
@@ -140,5 +96,7 @@ describe('ADMIN | Pages | Profile page | with SSO lock', () => {
     expect(queryByLabelText('Password')).not.toBeInTheDocument();
 
     expect(queryByLabelText('Password confirmation')).not.toBeInTheDocument();
+
+    server.resetHandlers();
   });
 });
