+++ packages/core/admin/admin/src/pages/SettingsPage/pages/ApplicationInfosPage/tests/index.test.js
@@ -1,51 +1,10 @@
 import React from 'react';
 
-import { fixtures } from '@strapi/admin-test-utils';
-import { lightTheme, ThemeProvider } from '@strapi/design-system';
-import { TrackingProvider, useAppInfo, useRBAC } from '@strapi/helper-plugin';
-import { fireEvent, render, waitFor, waitForElementToBeRemoved } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import { rest } from 'msw';
-import { setupServer } from 'msw/node';
-import { IntlProvider } from 'react-intl';
-import { QueryClient, QueryClientProvider } from 'react-query';
-import { Provider } from 'react-redux';
-import { createStore } from 'redux';
+import { useAppInfo, useRBAC } from '@strapi/helper-plugin';
+import { fireEvent, waitForElementToBeRemoved } from '@testing-library/react';
+import { render, waitFor } from '@tests/utils';
 
 import ApplicationInfosPage from '../index';
-
-const handlers = [
-  rest.get('*/project-settings', (req, res, ctx) => {
-    return res(
-      ctx.json({
-        menuLogo: {
-          ext: '.svg',
-          height: 256,
-          name: 'michka.svg',
-          size: 1.3,
-          url: '/uploads/michka.svg',
-          width: 256,
-        },
-      })
-    );
-  }),
-  rest.post('*/project-settings', (req, res, ctx) => {
-    return res(
-      ctx.json({
-        menuLogo: {
-          ext: '.svg',
-          height: 256,
-          name: 'michka.svg',
-          size: 1.3,
-          url: '/uploads/michka.svg',
-          width: 256,
-        },
-      })
-    );
-  }),
-];
-
-const server = setupServer(...handlers);
 
 const updateProjectSettingsSpy = jest.fn();
 
@@ -54,7 +13,6 @@ jest.mock('@strapi/helper-plugin', () => ({
   // eslint-disable-next-line
   CheckPermissions: ({ children }) => <div>{children}</div>,
   useAppInfo: jest.fn(() => ({ shouldUpdateStrapi: false, latestStrapiReleaseTag: 'v3.6.8' })),
-  useNotification: jest.fn(() => jest.fn()),
   useRBAC: jest.fn(() => ({ allowedActions: { canRead: true, canUpdate: true } })),
 }));
 
@@ -68,52 +26,13 @@ jest.mock('../../../../../hooks', () => ({
   })),
 }));
 
-const setup = (props) => ({
-  ...render(<ApplicationInfosPage {...props} />, {
-    wrapper({ children }) {
-      const client = new QueryClient({
-        defaultOptions: {
-          queries: {
-            retry: false,
-          },
-        },
-      });
-
-      return (
-        <Provider
-          store={createStore((state) => state, {
-            admin_app: { permissions: fixtures.permissions.app },
-          })}
-        >
-          <QueryClientProvider client={client}>
-            <TrackingProvider>
-              <ThemeProvider theme={lightTheme}>
-                <IntlProvider locale="en" messages={{}} textComponent="span">
-                  {children}
-                </IntlProvider>
-              </ThemeProvider>
-            </TrackingProvider>
-          </QueryClientProvider>
-        </Provider>
-      );
-    },
-  }),
-
-  user: userEvent.setup(),
-});
-
 describe('Application page', () => {
-  beforeAll(() => server.listen());
-
   afterEach(() => {
-    server.resetHandlers();
     jest.clearAllMocks();
   });
 
-  afterAll(() => server.close());
-
   it('should not display link upgrade version if not necessary', async () => {
-    const { queryByText } = setup();
+    const { queryByText } = render(<ApplicationInfosPage />);
 
     await waitForElementToBeRemoved(() => queryByText('Loading'));
 
@@ -127,7 +46,7 @@ describe('Application page', () => {
       strapiVersion: '4.0.0',
     });
 
-    const { getByText, queryByText } = setup();
+    const { getByText, queryByText } = render(<ApplicationInfosPage />);
 
     await waitForElementToBeRemoved(() => queryByText('Loading'));
 
@@ -136,7 +55,7 @@ describe('Application page', () => {
   });
 
   it('should render logo input if read permissions', async () => {
-    const { queryByText } = setup();
+    const { queryByText } = render(<ApplicationInfosPage />);
 
     await waitForElementToBeRemoved(() => queryByText('Loading'));
 
@@ -147,13 +66,13 @@ describe('Application page', () => {
     useRBAC.mockImplementationOnce(() => ({
       allowedActions: { canRead: false, canUpdate: false },
     }));
-    const { queryByText } = setup();
+    const { queryByText } = render(<ApplicationInfosPage />);
 
     expect(queryByText('Menu logo')).not.toBeInTheDocument();
   });
 
   it('should render save button if update permissions', async () => {
-    const { queryByText } = setup();
+    const { queryByText } = render(<ApplicationInfosPage />);
 
     await waitForElementToBeRemoved(() => queryByText('Loading'));
 
@@ -163,7 +82,7 @@ describe('Application page', () => {
   it('should not render save button if no update permissions', async () => {
     useRBAC.mockReturnValue({ allowedActions: { canRead: true, canUpdate: false } });
 
-    const { queryByText } = setup();
+    const { queryByText } = render(<ApplicationInfosPage />);
 
     await waitForElementToBeRemoved(() => queryByText('Loading'));
 
@@ -173,7 +92,7 @@ describe('Application page', () => {
   it('should update project settings on save', async () => {
     useRBAC.mockReturnValue({ allowedActions: { canRead: true, canUpdate: true } });
 
-    const { getByRole, queryByText } = setup();
+    const { getByRole, queryByText } = render(<ApplicationInfosPage />);
 
     await waitForElementToBeRemoved(() => queryByText('Loading'));
