+++ packages/core/admin/ee/admin/content-manager/pages/EditView/InformationBox/components/AssigneeSelect/tests/AssigneeSelect.test.js
@@ -1,98 +1,16 @@
 import React from 'react';
 
-import { fixtures } from '@strapi/admin-test-utils';
-import { ThemeProvider, lightTheme } from '@strapi/design-system';
-import { RBACContext, useCMEditViewDataManager } from '@strapi/helper-plugin';
-import { render, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { useCMEditViewDataManager } from '@strapi/helper-plugin';
+import { render, waitFor, server } from '@tests/utils';
 import { rest } from 'msw';
-import { setupServer } from 'msw/node';
-import { IntlProvider } from 'react-intl';
-import { QueryClientProvider, QueryClient } from 'react-query';
-import { Provider } from 'react-redux';
-import { createStore } from 'redux';
 
 import { ASSIGNEE_ATTRIBUTE_NAME } from '../../../constants';
 import { AssigneeSelect } from '../AssigneeSelect';
 
-const server = setupServer(
-  rest.get('*/users', (req, res, ctx) =>
-    res(
-      ctx.json({
-        data: {
-          results: [
-            {
-              id: 1,
-              firstname: 'Firstname 1',
-              lastname: 'Lastname 1',
-            },
-
-            {
-              id: 2,
-              firstname: 'Firstname 2',
-              username: 'jumbojosh',
-            },
-          ],
-        },
-      })
-    )
-  )
-);
-
 jest.mock('@strapi/helper-plugin', () => ({
   ...jest.requireActual('@strapi/helper-plugin'),
   useCMEditViewDataManager: jest.fn(),
-  useNotification: jest.fn(() => ({
-    toggleNotification: jest.fn(),
-  })),
 }));
-
-const setup = (props) => {
-  return {
-    user: userEvent.setup(),
-    ...render(<AssigneeSelect {...props} />, {
-      wrapper({ children }) {
-        const store = createStore((state = {}) => state, {
-          admin_app: {
-            permissions: fixtures.permissions.app,
-          },
-        });
-        const queryClient = new QueryClient({
-          defaultOptions: {
-            queries: {
-              retry: false,
-            },
-          },
-        });
-
-        return (
-          <RBACContext.Provider
-            value={{
-              allPermissions: [
-                ...fixtures.permissions.allPermissions,
-                {
-                  id: 314,
-                  action: 'admin::users.read',
-                  subject: null,
-                  properties: {},
-                  conditions: [],
-                },
-              ],
-            }}
-          >
-            <Provider store={store}>
-              <QueryClientProvider client={queryClient}>
-                <IntlProvider locale="en" defaultLocale="en">
-                  <ThemeProvider theme={lightTheme}>{children}</ThemeProvider>
-                </IntlProvider>
-              </QueryClientProvider>
-            </Provider>
-          </RBACContext.Provider>
-        );
-      },
-    }),
-  };
-};
 
 useCMEditViewDataManager.mockReturnValue({
   initialData: {
@@ -103,22 +21,17 @@ useCMEditViewDataManager.mockReturnValue({
 
 describe('EE | Content Manager | EditView | InformationBox | AssigneeSelect', () => {
   beforeAll(() => {
-    server.listen();
     jest.clearAllMocks();
   });
 
-  afterAll(() => {
-    server.close();
-  });
-
   it('renders a select with users, none is selected', async () => {
-    const { getByRole, getByText, queryByText, user } = setup();
+    const { getByRole, getByText, queryByText, user } = render(<AssigneeSelect />);
 
-    await waitFor(() => expect(queryByText('Firstname 1 Lastname 1')).not.toBeInTheDocument());
+    await waitFor(() => expect(queryByText('John Doe')).not.toBeInTheDocument());
 
     await user.click(getByRole('combobox'));
 
-    await waitFor(() => expect(getByText('Firstname 1 Lastname 1')).toBeInTheDocument());
+    await waitFor(() => expect(getByText('John Doe')).toBeInTheDocument());
   });
 
   it('renders a select with users, first user is selected', async () => {
@@ -126,16 +39,16 @@ describe('EE | Content Manager | EditView | InformationBox | AssigneeSelect', ()
       initialData: {
         [ASSIGNEE_ATTRIBUTE_NAME]: {
           id: 1,
-          firstname: 'Firstname 1',
-          lastname: 'Lastname 1',
+          firstname: 'John',
+          lastname: 'Doe',
         },
       },
       layout: { uid: 'api::articles:articles' },
     });
 
-    const { queryByRole } = setup();
+    const { queryByRole } = render(<AssigneeSelect />);
 
-    await waitFor(() => expect(queryByRole('combobox')).toHaveValue('Firstname 1 Lastname 1'));
+    await waitFor(() => expect(queryByRole('combobox')).toHaveValue('John Doe'));
   });
 
   it('renders a disabled select when there are no users to select', async () => {
@@ -143,15 +56,15 @@ describe('EE | Content Manager | EditView | InformationBox | AssigneeSelect', ()
       initialData: {
         [ASSIGNEE_ATTRIBUTE_NAME]: {
           id: 1,
-          firstname: 'Firstname 1',
-          lastname: 'Lastname 1',
+          firstname: 'John',
+          lastname: 'Doe',
         },
       },
       layout: { uid: 'api::articles:articles' },
     });
 
     server.use(
-      rest.get('*/users', (req, res, ctx) => {
+      rest.get('/admin/users', (req, res, ctx) => {
         return res.once(
           ctx.json({
             data: {
@@ -162,9 +75,11 @@ describe('EE | Content Manager | EditView | InformationBox | AssigneeSelect', ()
       })
     );
 
-    const { queryByRole } = setup();
+    const { queryByRole } = render(<AssigneeSelect />);
 
     await waitFor(() => expect(queryByRole('combobox')).toHaveAttribute('aria-disabled', 'true'));
+
+    server.resetHandlers();
   });
 
   it('renders an error message, when fetching user fails', async () => {
@@ -173,7 +88,7 @@ describe('EE | Content Manager | EditView | InformationBox | AssigneeSelect', ()
     console.error = jest.fn();
 
     server.use(
-      rest.get('*/users', (req, res, ctx) => {
+      rest.get('/admin/users', (req, res, ctx) => {
         return res.once(
           ctx.status(500),
           ctx.json({
@@ -187,13 +102,15 @@ describe('EE | Content Manager | EditView | InformationBox | AssigneeSelect', ()
       })
     );
 
-    const { getByText } = setup();
+    const { getByText } = render(<AssigneeSelect />);
 
     await waitFor(() =>
       expect(getByText('An error occurred while fetching users')).toBeInTheDocument()
     );
 
     console.error = origConsoleError;
+
+    server.resetHandlers();
   });
 
   it('renders an error message, when the assignee update fails', async () => {
@@ -210,29 +127,34 @@ describe('EE | Content Manager | EditView | InformationBox | AssigneeSelect', ()
     console.error = jest.fn();
 
     server.use(
-      rest.put('*/content-manager/collection-types/:uid/:entityId/assignee', (req, res, ctx) => {
-        return res.once(
-          ctx.status(500),
-          ctx.json({
-            data: {
-              error: {
-                message: 'Server side error message',
+      rest.put(
+        '/admin/content-manager/collection-types/:contentType/:id/assignee',
+        (req, res, ctx) => {
+          return res.once(
+            ctx.status(500),
+            ctx.json({
+              data: {
+                error: {
+                  message: 'Server side error message',
+                },
               },
-            },
-          })
-        );
-      })
+            })
+          );
+        }
+      )
     );
 
-    const { getByRole, getByText, user } = setup();
+    const { getByRole, getByText, user } = render(<AssigneeSelect />);
 
     await user.click(getByRole('combobox'));
-    await user.click(getByText('Firstname 1 Lastname 1'));
+    await user.click(getByText('John Doe'));
 
     await waitFor(() =>
       expect(getByText('Request failed with status code 500')).toBeInTheDocument()
     );
 
     console.error = origConsoleError;
+
+    server.resetHandlers();
   });
 });
