+++ packages/core/admin/package.json
@@ -135,6 +135,7 @@
     "yup": "0.32.9"
   },
   "devDependencies": {
+    "@strapi/admin-test-utils": "4.14.3",
     "@testing-library/dom": "9.2.0",
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
